+++ rollup.config.js
@@ -2,8 +2,8 @@ import alias from '@rollup/plugin-alias';
 import typescript from 'rollup-plugin-typescript2';
 import { terser } from 'rollup-plugin-terser';
 import json from '@rollup/plugin-json';
-// import nodeResolve from '@rollup/plugin-node-resolve';
-// import commonjs from '@rollup/plugin-commonjs';
+import nodeResolve from '@rollup/plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
 import replace from '@rollup/plugin-replace';
 import copy from 'rollup-plugin-copy';
 
@@ -12,18 +12,18 @@ import path from 'path';
 export default {
     input: 'src/index.ts',
     output: [
-        {
-            dir: 'dist/esm',
-            format: 'esm',
-            sourcemap: true,
-            compact: true,
-        },
         // {
-        //     dir: 'dist/cjs',
-        //     format: 'cjs',
+        //     dir: 'dist/esm',
+        //     format: 'esm',
         //     sourcemap: true,
         //     compact: true,
         // },
+        {
+            dir: 'dist/cjs',
+            format: 'cjs',
+            sourcemap: true,
+            compact: true,
+        },
     ],
     external: ['fs', 'path', 'https', 'url', 'crypto', 'bwip-js', 'xsd-schema-validator', 'pdfkit', 'pem'],
     plugins: [
@@ -37,8 +37,8 @@ export default {
             ],
         }),
         json(),
-        // nodeResolve(),
-        // commonjs(),
+        nodeResolve(),
+        commonjs(),
         typescript(),
         replace({
             'process.env.NODE_ENV': JSON.stringify('production'), // Substitui NODE_ENV por 'production'
@@ -47,9 +47,9 @@ export default {
         copy({
             targets: [
                 // ESM
-                { src: 'src/assets/*', dest: 'dist/esm/assets' },
-                { src: 'src/certs/*', dest: 'dist/esm/certs' },
-                { src: 'src/schemas/*', dest: 'dist/esm/schemas' },
+                // { src: 'src/assets/*', dest: 'dist/esm/assets' },
+                // { src: 'src/certs/*', dest: 'dist/esm/certs' },
+                // { src: 'src/schemas/*', dest: 'dist/esm/schemas' },
                 // CJS
                 { src: 'src/assets/*', dest: 'dist/cjs/assets' },
                 { src: 'src/certs/*', dest: 'dist/cjs/certs' },