+++ src/utils/Utility.ts
@@ -270,13 +270,11 @@ class Utility {
 
                 xsdValidator.validateXML(xml, xsdData, (err, validationResult) => {
                     if (err) {
-                        console.log({err})
                         reject({
                             success: false,
                             message: this.formatErrorMessage(err.message),
                         });
                     } else if (!validationResult.valid) {
-                        console.log({validationResult})
                         reject({
                             success: false,
                             message: this.formatErrorMessage(validationResult.messages[0]),